refactor: ♻️ merge entity-error and entity-attribute-error
