--- npm-packages/common-error-exception/src/lib/error-codes/domain/entity-attribute-error.type.ts
@@ -1,6 +1,0 @@
-import { ErrorObject } from '../shared/error-object.type';
-
-export type EntityAttributeError<T extends string | number | symbol> = Record<
-  T,
-  { [key: string]: ErrorObject }
->;
