+++ npm-packages/common-error-exception/src/lib/error-codes/domain/entity-error.type.ts
@@ -1,3 +1,8 @@
 import { ErrorObject } from "../shared/error-object.type";
 
 export type EntityError = { [key: string]: ErrorObject };
+
+export type EntityAttributeError<T extends string | number | symbol> = Record<
+  T,
+  { [key: string]: ErrorObject }
+>;
