+++ npm-packages/common-error-exception/src/lib/error-codes/shared/constants/entity-error-type.constant.ts
@@ -5,3 +5,13 @@ export const entityErrorType = {
 
 export type EntityErrorType =
   (typeof entityErrorType)[keyof typeof entityErrorType];
+
+export const entityAttributeErrorType = {
+  REQUIRED: 'REQUIRED',
+  INVALID_FORMAT: 'INVALID_FORMAT',
+  INVALID_VALUE: 'INVALID_VALUE',
+  LENGTH_MAX_EXCEEDED: 'LENGTH_MAX_EXCEEDED',
+} as const satisfies { [key: string]: string };
+
+export type EntityAttributeErrorType =
+  (typeof entityAttributeErrorType)[keyof typeof entityAttributeErrorType];
